Refactor TypeCastInsn class
