+++ compiler/BIR.h
@@ -89,26 +89,6 @@ public:
   void translate(LLVMModuleRef &modRef) final;
 };
 
-class TypeCastInsn : public NonTerminatorInsn {
-private:
-  Operand *rhsOp;
-  TypeDecl *typeDecl;
-  bool checkTypes;
-
-public:
-  TypeCastInsn() = default;
-  TypeCastInsn(Location *pos, InstructionKind kind, Operand *lOp, Operand *rOp,
-               TypeDecl *tDecl, bool checkTypes);
-  ~TypeCastInsn() = default;
-
-  Operand *getRhsOp() { return rhsOp; }
-  TypeDecl *getTypeDecl() { return typeDecl; }
-  bool mustCheckTypes() { return checkTypes; }
-  void setRhsOp(Operand *op) { rhsOp = op; }
-  void setTypeDecl(TypeDecl *tDecl) { typeDecl = tDecl; }
-  void setTypesChecking(bool checktypes) { checkTypes = checktypes; }
-  void translate(LLVMModuleRef &modRef) final;
-};
