+++ compiler/include/BalTypeCastInsn.h
@@ -1,0 +1,35 @@
+#ifndef __BALTYPECASTINSN__H__
+#define __BALTYPECASTINSN__H__
+
+#include "BalNonTerminatorInsn.h"
+
+namespace nballerina {
+
+// Forward Declare
+class TypeDecl;
+class Operand;
+
+class TypeCastInsn : public NonTerminatorInsn {
+private:
+  Operand *rhsOp;
+  TypeDecl *typeDecl;
+  bool checkTypes;
+
+public:
+  TypeCastInsn() = default;
+  TypeCastInsn(Location *pos, InstructionKind kind, Operand *lOp, Operand *rOp,
+               TypeDecl *tDecl, bool checkTypes);
+  ~TypeCastInsn() = default;
+
+  Operand *getRhsOp() { return rhsOp; }
+  TypeDecl *getTypeDecl() { return typeDecl; }
+  bool mustCheckTypes() { return checkTypes; }
+  void setRhsOp(Operand *op) { rhsOp = op; }
+  void setTypeDecl(TypeDecl *tDecl) { typeDecl = tDecl; }
+  void setTypesChecking(bool checktypes) { checkTypes = checktypes; }
+  void translate(LLVMModuleRef &modRef) final;
+};
+
+} // namespace nballerina
+
+#endif //!__BALTYPECASTINSN__H__
